interfaz de registro y listado tipo dashboard para ventas

+++ jama/src/main/java/com/cibertec/jama/repositories/menu/MenuSkuRepository.java
@@ -4,5 +4,6 @@ import com.cibertec.jama.entities.menu.MenuSku;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuSkuRepository  extends JpaRepository<MenuSku, Long> {
+public interface MenuSkuRepository  extends JpaRepository<MenuSku, Integer> {
+
 }
